added the trees back

+++ crates/rb_generation/src/trees/acacia.rs
@@ -1,9 +1,9 @@
-use rb_world::{BlockPos, VoxelWorld};
-use rb_block::Block;
 use super::utils::leaf_disk;
+use rb_block::Block;
+use rb_world::{BlockPos, VoxelWorld};
 
-pub fn grow_acacia(world: &VoxelWorld, pos: BlockPos, _seed: i32, dist: f32) {
-    let height = 10-(dist*7.) as i32;
+pub fn grow_acacia(world: &VoxelWorld, pos: BlockPos, _seed: u64, dist: f32) {
+    let height = 10 - (dist * 7.) as i32;
     let mut pos = pos;
     for _ in 0..height {
         world.set_block(pos, Block::AcaciaLog);
@@ -13,11 +13,11 @@ pub fn grow_acacia(world: &VoxelWorld, pos: BlockPos, _seed: i32, dist: f32) {
     pos.y -= 1;
     leaf_disk(world, pos, 1, Block::AcaciaLeaves);
     pos.y += 1;
-    leaf_disk(world, pos, height as u32-3, Block::AcaciaLeaves);
+    leaf_disk(world, pos, height as u32 - 3, Block::AcaciaLeaves);
     pos.y += 1;
-    leaf_disk(world, pos, height as u32-4, Block::AcaciaLeaves);
+    leaf_disk(world, pos, height as u32 - 4, Block::AcaciaLeaves);
     if height > 6 {
         pos.y += 1;
-        leaf_disk(world, pos, height as u32-5, Block::AcaciaLeaves);
+        leaf_disk(world, pos, height as u32 - 5, Block::AcaciaLeaves);
     }
 }
